requirements.txt
fastapi==0.116.1
mysql_connector_repackaged==0.3.1
pandas==2.3.1
pydantic==2.11.7
Requests==2.32.4
streamlit==1.47.1
